chore: переход на fsd. widgets
