+++ src/components/Articles/index.tsx
@@ -1,7 +1,7 @@
 import cn from "classnames";
 import { ReactNode } from "react";
-import IconedButton from "../shared/IconedButton";
-import Title from '../Title';
+
+import { IconedButton, Title } from '../../widgets';
 import styles from "./index.module.css";
 export default function Articles({ children }: { children: ReactNode }) {
   return (
