--- src/components/shared/Button/button.props.ts
@@ -1,5 +1,0 @@
-export default interface ButtonProps {
-    className?: string;
-    onClick: (event: MouseEvent) => void;
-    children: React.ReactNode;
-}
