--- src/components/shared/IconedButton/button.props.ts
@@ -1,11 +1,0 @@
-import { ReactNode } from 'react';
-
-export default interface IconedButtonProps {
-    children: ReactNode,
-    iconUrl: string,
-    className?: string,
-    showInMobile?: boolean,
-    iconWidth?: number,
-    iconHeight?: number,
-    linkUrl?: string
-}
