+++ src/widgets/Button/button.props.ts
@@ -1,0 +1,10 @@
+import { MouseEventHandler } from 'react';
+
+/**
+ * Simple button. No links
+ */
+export default interface ButtonProps {
+    className?: string;
+    onClick: MouseEventHandler<HTMLButtonElement>;
+    children: React.ReactNode;
+}
