+++ src/widgets/index.tsx
@@ -1,0 +1,6 @@
+import Button from './Button';
+import IconedButton from './IconedButton';
+import Title from './Title';
+
+export { Button, IconedButton, Title };
+
